+++ experimental/teachable_machine.py
@@ -30,18 +30,8 @@ if cfg.video:
     fourcc = cv2.VideoWriter_fourcc(*'mp4v')
     out = cv2.VideoWriter(name, fourcc, cfg.fps, (cfg.w, cfg.h))
 
-def source_capture(source):
-    source = int(source) if source.isdigit() else source
-    cap = cv2.VideoCapture(source)
-
-    fourcc_cap = cv2.VideoWriter_fourcc(*'MJPG')
-    cap.set(cv2.CAP_PROP_FOURCC, fourcc_cap)
-    cap.set(cv2.CAP_PROP_FRAME_WIDTH, cfg.w)
-    cap.set(cv2.CAP_PROP_FRAME_HEIGHT, cfg.h)
-    return cap
-
 trackers = []
-cap = source_capture(sys.argv[1])
+cap = utils.source_capture(sys.argv[1])
 img = utils.img_proc()
 
 while True:
@@ -55,26 +45,7 @@ while True:
             bbox = utils.get_bbox(list(body.values()))
             bboxes.append((bbox, body))
 
-        track_boxes = [tracker.bbox for tracker in trackers]
-        matched, unmatched_trackers, unmatched_detections = utils.tracker_match(track_boxes, [b[0] for b in bboxes])
-
-        for idx, jdx in matched:
-            trackers[idx].set_bbox(bboxes[jdx][0])
-            trackers[idx].update_pose(bboxes[jdx][1])
-
-        for idx in unmatched_detections:
-            try:
-                trackers[idx].count += 1
-                if trackers[idx].count > trackers[idx].expiration:
-                    trackers.pop(idx)
-            except:
-                pass
-
-        for idx in unmatched_trackers:
-            p = person.PersonTracker()
-            p.set_bbox(bboxes[idx][0])
-            p.update_pose(bboxes[idx][1])
-            trackers.append(p)
+        trackers = utils.update_trackers(trackers, bboxes)
 
         pprint([(tracker.id, np.vstack(tracker.q)) for tracker in trackers])
 
@@ -97,7 +68,6 @@ while True:
 
             if cfg.log:
                 newFileWriter.writerow([tracker.activity] + list(np.hstack(list(tracker.q)[:cfg.window])))
-                    
 
         if cfg.video:
             for tracker in trackers:
